refactor: moved order mapping to caching function
