+++ packages/server/src/queries/complex/orderbooks/denoms.ts
@@ -1,4 +1,5 @@
-import { Chain } from "@osmosis-labs/types";
+import { AssetList, Chain } from "@osmosis-labs/types";
+import { getAssetFromAssetList } from "@osmosis-labs/utils";
 import cachified, { CacheEntry } from "cachified";
 import { LRUCache } from "lru-cache";
 
@@ -12,17 +13,30 @@ const orderbookDenomsCache = new LRUCache<string, CacheEntry>(
 export function getOrderbookDenoms({
   orderbookAddress,
   chainList,
+  assetLists,
 }: {
   orderbookAddress: string;
   chainList: Chain[];
+  assetLists: AssetList[];
 }) {
   return cachified({
     cache: orderbookDenomsCache,
     key: `orderbookDenoms-${orderbookAddress}`,
-    ttl: 1000 * 60 * 60 * 24 * 7, // 7 days
+    ttl: 1000 * 60 * 60 * 24 * 30, // 30 days
     getFreshValue: () =>
       queryOrderbookDenoms({ orderbookAddress, chainList }).then(
-        ({ data }) => data
+        ({ data: { quote_denom, base_denom } }) => {
+          const quoteAsset = getAssetFromAssetList({
+            coinMinimalDenom: quote_denom,
+            assetLists,
+          });
+          const baseAsset = getAssetFromAssetList({
+            coinMinimalDenom: base_denom,
+            assetLists,
+          });
+
+          return { quoteAsset, baseAsset };
+        }
       ),
   });
 }
