+++ packages/trpc/src/orderbook-router.ts
@@ -3,20 +3,13 @@ import { tickToPrice } from "@osmosis-labs/math";
 import {
   CursorPaginationSchema,
   getOrderbookActiveOrders,
-  getOrderbookDenoms,
   getOrderbookHistoricalOrders,
   getOrderbookMakerFee,
   getOrderbookPools,
   getOrderbookState,
-  getOrderbookTickState,
-  getOrderbookTickUnrealizedCancels,
-  HistoricalLimitOrder,
-  LimitOrder,
   maybeCachePaginatedItems,
 } from "@osmosis-labs/server";
-import { dayjs } from "@osmosis-labs/server/build/utils/dayjs";
-import { Chain } from "@osmosis-labs/types";
-import { getAssetFromAssetList } from "@osmosis-labs/utils";
+import { MappedLimitOrder } from "@osmosis-labs/server/build/queries/complex/orderbooks/types";
 import { z } from "zod";
 
 import { createTRPCRouter, publicProcedure } from "./api";
@@ -28,40 +21,6 @@ const GetInfiniteLimitOrdersInputSchema = CursorPaginationSchema.merge(
   })
 );
 
-export type OrderStatus =
-  | "open"
-  | "partiallyFilled"
-  | "filled"
-  | "fullyClaimed"
-  | "cancelled";
-
-export type MappedLimitOrder = Omit<
-  LimitOrder,
-  "quantity" | "placed_quantity" | "placed_at"
-> & {
-  quantity: number;
-  placed_quantity: number;
-  percentClaimed: Dec;
-  totalFilled: number;
-  percentFilled: Dec;
-  orderbookAddress: string;
-  price: Dec;
-  status: OrderStatus;
-  output: Dec;
-  quoteAsset: ReturnType<typeof getAssetFromAssetList>;
-  baseAsset: ReturnType<typeof getAssetFromAssetList>;
-  placed_at: number;
-};
-
-function mapOrderStatus(order: LimitOrder, percentFilled: Dec): OrderStatus {
-  const quantInt = parseInt(order.quantity);
-  if (quantInt === 0 || percentFilled.equals(new Dec(1))) return "filled";
-  if (percentFilled.isZero()) return "open";
-  if (percentFilled.lt(new Dec(1))) return "partiallyFilled";
-
-  return "open";
-}
-
 function defaultSortOrders(
   orderA: MappedLimitOrder,
   orderB: MappedLimitOrder
@@ -72,144 +31,6 @@ function defaultSortOrders(
   if (orderA.status === "filled") return -1;
   if (orderB.status === "filled") return 1;
   return orderB.placed_at - orderA.placed_at;
-}
-
-async function getTickInfoAndTransformOrders(
-  orderbookAddress: string,
-  orders: LimitOrder[],
-  chainList: Chain[],
-  quoteAsset: ReturnType<typeof getAssetFromAssetList>,
-  baseAsset: ReturnType<typeof getAssetFromAssetList>
-): Promise<MappedLimitOrder[]> {
-  const tickIds = [...new Set(orders.map((o) => o.tick_id))];
-  const tickStates = await getOrderbookTickState({
-    orderbookAddress,
-    chainList,
-    tickIds,
-  });
-  const unrealizedTickCancels = await getOrderbookTickUnrealizedCancels({
-    orderbookAddress,
-    chainList,
-    tickIds,
-  });
-
-  const fullTickState = tickStates.map(({ tick_id, tick_state }) => ({
-    tickId: tick_id,
-    tickState: tick_state,
-    unrealizedCancels: unrealizedTickCancels.find((c) => c.tick_id === tick_id),
-  }));
-
-  return orders.map((o) => {
-    const { tickState, unrealizedCancels } = fullTickState.find(
-      ({ tickId }) => tickId === o.tick_id
-    ) ?? { tickState: undefined, unrealizedCancels: undefined };
-
-    const quantity = parseInt(o.quantity);
-    const placedQuantity = parseInt(o.placed_quantity);
-
-    const percentClaimed = new Dec(
-      (placedQuantity - quantity) / placedQuantity
-    );
-
-    const normalizationFactor = new Dec(10).pow(
-      new Int((quoteAsset?.decimals ?? 0) - (baseAsset?.decimals ?? 0))
-    );
-    const [tickEtas, tickUnrealizedCancelled] =
-      o.order_direction === "bid"
-        ? [
-            parseInt(
-              tickState?.bid_values.effective_total_amount_swapped ?? "0"
-            ),
-            parseInt(
-              unrealizedCancels?.unrealized_cancels.bid_unrealized_cancels ??
-                "0"
-            ),
-          ]
-        : [
-            parseInt(
-              tickState?.ask_values.effective_total_amount_swapped ?? "0"
-            ),
-            parseInt(
-              unrealizedCancels?.unrealized_cancels.ask_unrealized_cancels ??
-                "0"
-            ),
-          ];
-    const tickTotalEtas = tickEtas + tickUnrealizedCancelled;
-    const totalFilled = Math.max(
-      tickTotalEtas - (parseInt(o.etas) - (placedQuantity - quantity)),
-      0
-    );
-    const percentFilled = new Dec(Math.min(totalFilled / placedQuantity, 1));
-    const price = tickToPrice(new Int(o.tick_id));
-    const status = mapOrderStatus(o, percentFilled);
-    const output =
-      o.order_direction === "bid"
-        ? new Dec(placedQuantity).quo(price)
-        : new Dec(placedQuantity).mul(price);
-    return {
-      ...o,
-      price: price.quo(normalizationFactor),
-      quantity,
-      placed_quantity: placedQuantity,
-      percentClaimed,
-      totalFilled,
-      percentFilled,
-      orderbookAddress,
-      status,
-      output,
-      quoteAsset,
-      baseAsset,
-      placed_at: dayjs(parseInt(o.placed_at) / 1_000).unix(),
-    };
-  });
-}
-
-function mapHistoricalToMapped(
-  historicalOrders: HistoricalLimitOrder[],
-  userAddress: string,
-  quoteAsset: ReturnType<typeof getAssetFromAssetList>,
-  baseAsset: ReturnType<typeof getAssetFromAssetList>
-): MappedLimitOrder[] {
-  return historicalOrders.map((o) => {
-    const quantityMin = parseInt(o.quantity);
-    const placedQuantityMin = parseInt(o.quantity);
-    const price = tickToPrice(new Int(o.tick_id));
-    const percentClaimed = new Dec(1);
-    const output =
-      o.order_direction === "bid"
-        ? new Dec(placedQuantityMin).quo(price)
-        : new Dec(placedQuantityMin).mul(price);
-
-    const normalizationFactor = new Dec(10).pow(
-      new Int((quoteAsset?.decimals ?? 0) - (baseAsset?.decimals ?? 0))
-    );
-    return {
-      quoteAsset,
-      baseAsset,
-      etas: "0",
-      order_direction: o.order_direction,
-      order_id: parseInt(o.order_id),
-      owner: userAddress,
-      placed_at:
-        dayjs(
-          o.place_timestamp && o.place_timestamp.length > 0
-            ? o.place_timestamp
-            : 0
-        ).unix() * 1000,
-      placed_quantity: parseInt(o.quantity),
-      placedQuantityMin,
-      quantityMin,
-      quantity: parseInt(o.quantity),
-      price: price.quo(normalizationFactor),
-      status: o.status as OrderStatus,
-      tick_id: parseInt(o.tick_id),
-      output,
-      percentClaimed,
-      percentFilled: new Dec(1),
-      totalFilled: parseInt(o.quantity),
-      orderbookAddress: o.contract,
-    };
-  });
 }
 
 export const orderbookRouter = createTRPCRouter({
@@ -225,7 +46,7 @@ export const orderbookRouter = createTRPCRouter({
         makerFee,
       };
     }),
-  getAllActiveOrders: publicProcedure
+  getAllOrders: publicProcedure
     .input(
       GetInfiniteLimitOrdersInputSchema.merge(
         z.object({ contractAddresses: z.array(z.string().startsWith("osmo")) })
@@ -240,54 +61,23 @@ export const orderbookRouter = createTRPCRouter({
 
           const historicalOrders = await getOrderbookHistoricalOrders({
             userOsmoAddress: input.userOsmoAddress,
+            assetLists: ctx.assetLists,
+            chainList: ctx.chainList,
           });
 
           const promises = contractAddresses.map(
             async (contractOsmoAddress: string) => {
-              const resp = await getOrderbookActiveOrders({
+              const orders = await getOrderbookActiveOrders({
                 orderbookAddress: contractOsmoAddress,
                 userOsmoAddress: userOsmoAddress,
                 chainList: ctx.chainList,
+                assetLists: ctx.assetLists,
               });
               const historicalOrdersForContract = historicalOrders.filter(
-                (o) => o.contract === contractOsmoAddress
+                (o) => o.orderbookAddress === contractOsmoAddress
               );
 
-              if (
-                resp.orders.length === 0 &&
-                historicalOrdersForContract.length === 0
-              )
-                return [];
-              const { base_denom, quote_denom } = await getOrderbookDenoms({
-                orderbookAddress: contractOsmoAddress,
-                chainList: ctx.chainList,
-              });
-              const quoteAsset = getAssetFromAssetList({
-                assetLists: ctx.assetLists,
-                coinMinimalDenom: quote_denom,
-              });
-
-              const baseAsset = getAssetFromAssetList({
-                assetLists: ctx.assetLists,
-                coinMinimalDenom: base_denom,
-              });
-
-              const mappedOrders = await getTickInfoAndTransformOrders(
-                contractOsmoAddress,
-                resp.orders,
-                ctx.chainList,
-                quoteAsset,
-                baseAsset
-              );
-
-              const mappedHistoricalOrders = mapHistoricalToMapped(
-                historicalOrdersForContract,
-                input.userOsmoAddress,
-                quoteAsset,
-                baseAsset
-              );
-
-              return [...mappedOrders, ...mappedHistoricalOrders];
+              return [...orders, ...historicalOrdersForContract];
             }
           );
           const ordersByContracts = await Promise.all(promises);
@@ -330,34 +120,16 @@ export const orderbookRouter = createTRPCRouter({
       const { contractAddresses, userOsmoAddress } = input;
       const promises = contractAddresses.map(
         async (contractOsmoAddress: string) => {
-          const resp = await getOrderbookActiveOrders({
+          const orders = await getOrderbookActiveOrders({
             orderbookAddress: contractOsmoAddress,
             userOsmoAddress: userOsmoAddress,
             chainList: ctx.chainList,
+            assetLists: ctx.assetLists,
           });
 
-          if (resp.orders.length === 0) return [];
-          const { base_denom, quote_denom } = await getOrderbookDenoms({
-            orderbookAddress: contractOsmoAddress,
-            chainList: ctx.chainList,
-          });
-          // TODO: Use actual quote denom here
-          const quoteAsset = getAssetFromAssetList({
-            assetLists: ctx.assetLists,
-            coinMinimalDenom: quote_denom,
-          });
-          const baseAsset = getAssetFromAssetList({
-            assetLists: ctx.assetLists,
-            coinMinimalDenom: base_denom,
-          });
-          const mappedOrders = await getTickInfoAndTransformOrders(
-            contractOsmoAddress,
-            resp.orders,
-            ctx.chainList,
-            quoteAsset,
-            baseAsset
-          );
-          return mappedOrders.filter((o) => o.percentFilled.gte(new Dec(1)));
+          if (orders.length === 0) return [];
+
+          return orders.filter((o) => o.percentFilled.gte(new Dec(1)));
         }
       );
       const ordersByContracts = await Promise.all(promises);
@@ -366,10 +138,12 @@ export const orderbookRouter = createTRPCRouter({
     }),
   getHistoricalOrders: publicProcedure
     .input(UserOsmoAddressSchema.required())
-    .query(async ({ input }) => {
+    .query(async ({ input, ctx }) => {
       const { userOsmoAddress } = input;
       const historicalOrders = await getOrderbookHistoricalOrders({
         userOsmoAddress,
+        assetLists: ctx.assetLists,
+        chainList: ctx.chainList,
       });
       return historicalOrders;
     }),
